refactor: move pool-detail components
