--- packages/web/src/components/pool/my-liquidity-details/MyLiquidityDetails.styles.ts
@@ -1,8 +1,0 @@
-import { css } from "@emotion/react";
-import mixins from "@styles/mixins";
-
-export const wrapper = () => css`
-  ${mixins.flexbox("column", "center", "center")};
-  width: 100%;
-  gap: 16px;
-`;
