+++ packages/web/src/views/pool/pool-detail/components/my-liquidity/my-detailed-position-card/PositionHistory.styles.ts
