+++ packages/web/src/views/pool/pool-detail/components/pool-pair-information/PoolPairInformation.styles.ts
