+++ packages/web/src/views/pool/pool-detail/components/pool-pair-information/pool-pair-info-content/PoolPairInfoContent.styles.ts
