+++ packages/web/src/views/pool/pool-detail/components/pool-pair-information/pool-pair-info-header/PoolPairInfoHeader.spec.tsx
@@ -6,7 +6,7 @@ import { PoolRepositoryMock } from "@repositories/pool";
 import GnoswapServiceProvider from "@providers/gnoswap-service-provider/GnoswapServiceProvider";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { AppRouterContext } from "next/dist/shared/lib/app-router-context";
-import { createMockRouter } from "../../../test-utils/createMockRouter";
+import { createMockRouter } from "../../../../../../test-utils/createMockRouter";
 
 const poolRepository = new PoolRepositoryMock();
